Almost completed Board class

+++ Curvefever/src/CurveFever/PackageS2C.java
@@ -4,10 +4,10 @@ public class PackageS2C implements java.io.Serializable{
     public CurvePoint[] CurvePoints;
     public int currentRound;
     public GameState gameState;
-    public int[] Scores;
+    public double[] Scores;
 
     public PackageS2C(int numOfPlayers) {
         CurvePoints = new CurvePoint[numOfPlayers];
-        Scores = new int[numOfPlayers];
+        Scores = new double[numOfPlayers];
     }
 }
